Move streaming API to different service class
ValintatulosService is too large.
No functional changes.
