+++ valinta-tulos-service/src/main/scala/fi/vm/sade/valintatulosservice/ValintatulosService.scala
@@ -173,7 +173,7 @@ class ValintatulosService(valinnantulosRepository: ValinnantulosRepository,
     }
   }
 
-  private def hakemustenTulosByHaku(hakuOid: HakuOid, haunVastaanotot: Option[Map[String,Set[VastaanottoRecord]]], checkJulkaisuAikaParametri: Boolean = true): Option[Iterator[Hakemuksentulos]] = {
+  def hakemustenTulosByHaku(hakuOid: HakuOid, haunVastaanotot: Option[Map[String,Set[VastaanottoRecord]]], checkJulkaisuAikaParametri: Boolean = true): Option[Iterator[Hakemuksentulos]] = {
     timed("Fetch hakemusten tulos for haku: " + hakuOid, 1000) (
       for {
         haku <- hakuService.getHaku(hakuOid).right.toOption
@@ -399,107 +399,18 @@ class ValintatulosService(valinnantulosRepository: ValinnantulosRepository,
     })
   }
 
-  def streamSijoittelunTulokset(hakuOid: HakuOid,
-                                sijoitteluajoId: String,
-                                hakukohdeOids: Set[HakukohdeOid],
-                                writeResult: HakijaDTO => Unit,
-                                vainMerkitsevaJono : Boolean): Unit = {
-    val haunVastaanototByHakijaOid = timed("Fetch haun vastaanotot for haku: " + hakuOid, 1000) {
-      virkailijaVastaanottoRepository.findHaunVastaanotot(hakuOid).groupBy(_.henkiloOid)
-    }
-    val hakemustenTulokset: Option[Iterator[Hakemuksentulos]] = Some(hakukohdeOids.map { hakukohdeOid =>
-      hakemustenTulosByHakukohde(hakuOid, hakukohdeOid, Some(haunVastaanototByHakijaOid)) match {
-        case Right(it) => it
-        case Left(e) => val msg = s"Could not retrieve results for hakukohde $hakukohdeOid of haku $hakuOid"
-          logger.error(msg)
-          throw new RuntimeException(msg)
-      }
-    }.reduce((i1, i2) => i1 ++ i2))
-    val hakutoiveidenTuloksetByHakemusOid: Map[HakemusOid, (String, List[Hakutoiveentulos])] =
-      timed(s"Find hakutoiveiden tulokset for ${hakukohdeOids.size} hakukohdes of haku $hakuOid", 1000) {
-        hakemustenTulokset match {
-          case Some(hakemustenTulosIterator) => hakemustenTulosIterator.map(h => (h.hakemusOid, (h.hakijaOid, h.hakutoiveet))).toMap
-          case None => Map()
-        }
-      }
-    logger.info(s"Found ${hakutoiveidenTuloksetByHakemusOid.keySet.size} hakemus objects for sijoitteluajo $sijoitteluajoId " +
-      s"of ${hakukohdeOids.size} hakukohdes of haku $hakuOid")
-
-    streamSijoittelunTulokset(hakutoiveidenTuloksetByHakemusOid, hakuOid, sijoitteluajoId, writeResult, vainMerkitsevaJono)
-  }
-
-
-  def streamSijoittelunTulokset(hakuOid: HakuOid, sijoitteluajoId: String, writeResult: HakijaDTO => Unit, vainMerkitsevaJono : Boolean): Unit = {
-    val haunVastaanototByHakijaOid = timed("Fetch haun vastaanotot for haku: " + hakuOid, 1000) {
-      virkailijaVastaanottoRepository.findHaunVastaanotot(hakuOid).groupBy(_.henkiloOid)
-    }
-    val hakemustenTulokset: Option[Iterator[Hakemuksentulos]] = hakemustenTulosByHaku(hakuOid, Some(haunVastaanototByHakijaOid))
-    val hakutoiveidenTuloksetByHakemusOid: Map[HakemusOid, (String, List[Hakutoiveentulos])] = timed(s"Find hakutoiveiden tulokset for haku $hakuOid", 1000) {
-      hakemustenTulokset match {
-        case Some(hakemustenTulosIterator) => hakemustenTulosIterator.map(h => (h.hakemusOid, (h.hakijaOid, h.hakutoiveet))).toMap
-        case None => Map()
-      }
-    }
-    logger.info(s"Found ${hakutoiveidenTuloksetByHakemusOid.keySet.size} hakemus objects for sijoitteluajo $sijoitteluajoId of haku $hakuOid")
-
-    streamSijoittelunTulokset(hakutoiveidenTuloksetByHakemusOid, hakuOid, sijoitteluajoId, writeResult, vainMerkitsevaJono)
-  }
-
-  private def streamSijoittelunTulokset(tuloksetByHakemusOid: Map[HakemusOid, (String, List[Hakutoiveentulos])],
-                                        hakuOid: HakuOid,
-                                        sijoitteluajoId: String,
-                                        writeResult: HakijaDTO => Unit,
-                                        vainMerkitsevaJono: Boolean): Unit = {
-    def processTulos(hakijaDto: HakijaDTO, hakijaOid: String, hakutoiveidenTulokset: List[Hakutoiveentulos]): Unit = {
-      hakijaDto.setHakijaOid(hakijaOid)
-      if (vainMerkitsevaJono) {
-        populateMerkitsevatJonot(hakijaDto, hakutoiveidenTulokset)
-      } else {
-        populateVastaanottotieto(hakijaDto, hakutoiveidenTulokset)
-      }
-      writeResult(hakijaDto)
-    }
-
-    try {
-      hakijaDTOClient.processSijoittelunTulokset(hakuOid, sijoitteluajoId, { hakijaDto: HakijaDTO =>
-        tuloksetByHakemusOid.get(HakemusOid(hakijaDto.getHakemusOid)) match {
-          case Some((hakijaOid, hakutoiveidenTulokset)) => processTulos(hakijaDto, hakijaOid, hakutoiveidenTulokset)
-          case None => crashOrLog(s"Hakemus ${hakijaDto.getHakemusOid} not found in hakemusten tulokset for haku $hakuOid")
-        }
-      })
-    } catch {
-      case e: Exception =>
-        logger.error(s"Sijoitteluajon $sijoitteluajoId hakemuksia ei saatu palautettua haulle $hakuOid", e)
-        throw e
-    }
-  }
-
-  private def copyVastaanottotieto(hakutoiveDto: HakutoiveDTO, hakutoiveenTulos: Hakutoiveentulos) = {
+  def copyVastaanottotieto(hakutoiveDto: HakutoiveDTO, hakutoiveenTulos: Hakutoiveentulos) = {
     hakutoiveDto.setVastaanottotieto(fi.vm.sade.sijoittelu.tulos.dto.ValintatuloksenTila.valueOf(hakutoiveenTulos.vastaanottotila.toString))
     hakutoiveDto.getHakutoiveenValintatapajonot.asScala.foreach(_.setTilanKuvaukset(hakutoiveenTulos.tilanKuvaukset.asJava))
   }
 
-  private def populateVastaanottotieto(hakijaDto: HakijaDTO, hakemuksenHakutoiveidenTuloksetVastaanottotiedonKanssa: List[Hakutoiveentulos]): Unit = {
+  def populateVastaanottotieto(hakijaDto: HakijaDTO, hakemuksenHakutoiveidenTuloksetVastaanottotiedonKanssa: List[Hakutoiveentulos]): Unit = {
     hakijaDto.getHakutoiveet.asScala.foreach(palautettavaHakutoiveDto =>
       hakemuksenHakutoiveidenTuloksetVastaanottotiedonKanssa.find(_.hakukohdeOid.toString == palautettavaHakutoiveDto.getHakukohdeOid) match {
         case Some(hakutoiveenOikeaTulos) => copyVastaanottotieto(palautettavaHakutoiveDto, hakutoiveenOikeaTulos)
         case None => palautettavaHakutoiveDto.setVastaanottotieto(dto.ValintatuloksenTila.KESKEN)
       }
     )
-  }
-
-  private def populateMerkitsevatJonot(hakijaDto: HakijaDTO, hakemuksenHakutoiveidenTuloksetVastaanottotiedonKanssa: List[Hakutoiveentulos]) = {
-    hakijaDto.getHakutoiveet.asScala.foreach(hakutoiveDto => {
-      hakemuksenHakutoiveidenTuloksetVastaanottotiedonKanssa.find(_.hakukohdeOid.toString == hakutoiveDto.getHakukohdeOid) match {
-        case Some(hakutoiveenTulos: Hakutoiveentulos) => {
-          hakutoiveDto.setHakutoiveenValintatapajonot(
-            hakutoiveDto.getHakutoiveenValintatapajonot.asScala.filter(_.getValintatapajonoOid == hakutoiveenTulos.valintatapajonoOid.toString).asJava
-          )
-          copyVastaanottotieto(hakutoiveDto, hakutoiveenTulos)
-        }
-        case None => hakutoiveDto.setHakutoiveenValintatapajonot(List().asJava)
-      }
-    })
   }
 
   private def mapHakemustenTuloksetByHakemusOid(hakemustenTulokset:Iterator[Hakemuksentulos]):Map[HakemusOid, Hakemuksentulos] = {
@@ -890,7 +801,7 @@ class ValintatulosService(valinnantulosRepository: ValinnantulosRepository,
     }
   }
 
-  private def crashOrLog[T](msg: String): Option[T] = {
+  def crashOrLog[T](msg: String): Option[T] = {
     if (appConfig.settings.lenientSijoitteluntuloksetParsing) {
       logger.warn(msg)
       None
